test: 테스트 코드 작성
- 테스트 코드를 열심히 작성하였습니다

+++ winwin-be-domain/src/main/java/com/dpm/winwin/domain/dto/post/MyPagePostDto.java
@@ -3,8 +3,12 @@ package com.dpm.winwin.domain.dto.post;
 import com.querydsl.core.annotations.QueryProjection;
 import java.util.List;
 
-public record MyPagePostDto(Long id, String title, String subCategory, boolean isShare, List<String> takenTalents,
-                            Long likes) {
+public record MyPagePostDto(Long id,
+                            String title,
+                            String subCategory,
+                            boolean isShare,
+                            List<String> takenTalents,
+                            int likes) {
 
 
     @QueryProjection
